requirements.txt
gradio
PyPDF2
openai
Pillow
pytesseract
pymupdf
easyocr
torch
opencv-python
google-genai
python-dotenv
google.generativeai